+++ extension/package.json
@@ -7,8 +7,9 @@
     "pnpm": "^9.0.0"
   },
   "scripts": {
-    "dev": "node build.js dev && vite",
-    "build": "node build.js prod && vite build"
+    "dev": "../scripts/tsnode build.ts dev && vite",
+    "build": "../scripts/tsnode build.ts prod && vite build",
+    "test": "../scripts/tsnode build.ts dev"
   },
   "dependencies": {
     "@crxjs/vite-plugin": "2.0.0-beta.23",
